Bugfixes and 64bit compatibilty patch that was submitted by Christoph

+++ NSString+iMedia.h
@@ -66,7 +66,7 @@
 
 - (NSString *)exifDateToLocalizedDisplayDate;
 
-+ (NSString *)stringFromStarRating:(unsigned int)aRating;
++ (NSString *)stringFromStarRating:(NSUInteger)aRating;
 
 - (NSComparisonResult)finderCompare:(NSString *)aString;
